refactor(dynamodb): split 7,935-line service.rs into sub-modules
Split monolithic service.rs into 7 files following the SSM pattern:
- mod.rs: struct, dispatch, expression evaluation, helpers, tests (4,829 lines)
- tables.rs: table CRUD, TTL, tags, backups (1,378 lines)
- batch.rs: batch/transact operations, PartiQL (564 lines)
- items.rs: PutItem, GetItem, DeleteItem, UpdateItem (407 lines)
- queries.rs: Query, Scan (328 lines)
- streams.rs: endpoints, limits, Kinesis destinations (299 lines)
- global_tables.rs: global table operations (281 lines)
