+++ crates/fakecloud-dynamodb/src/service/mod.rs
@@ -1,24 +1,27 @@
+mod batch;
+mod global_tables;
+mod items;
+mod queries;
+mod streams;
+mod tables;
+
 use std::collections::HashMap;
 use std::sync::Arc;
 
 use async_trait::async_trait;
 use base64::Engine;
-use chrono::Utc;
 use http::StatusCode;
-use parking_lot::RwLock;
 use serde_json::{json, Value};
 
 use fakecloud_core::delivery::DeliveryBus;
 use fakecloud_core::service::{AwsRequest, AwsResponse, AwsService, AwsServiceError};
-use fakecloud_core::validation::*;
 
 use fakecloud_s3::state::SharedS3State;
 
 use crate::state::{
-    attribute_type_and_value, AttributeDefinition, AttributeValue, BackupDescription, DynamoTable,
-    ExportDescription, GlobalSecondaryIndex, GlobalTableDescription, ImportDescription,
-    KeySchemaElement, KinesisDestination, LocalSecondaryIndex, Projection, ProvisionedThroughput,
-    ReplicaDescription, SharedDynamoDbState,
+    attribute_type_and_value, AttributeDefinition, AttributeValue, DynamoTable,
+    GlobalSecondaryIndex, KeySchemaElement, LocalSecondaryIndex, Projection, ProvisionedThroughput,
+    SharedDynamoDbState,
 };
 
 pub struct DynamoDbService {
@@ -114,3114 +117,6 @@ impl DynamoDbService {
 
     fn ok_json(body: Value) -> Result<AwsResponse, AwsServiceError> {
         Ok(AwsResponse::ok_json(body))
-    }
-
-    // ── Table Operations ────────────────────────────────────────────────
-
-    fn create_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-
-        let table_name = body["TableName"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "TableName is required",
-                )
-            })?
-            .to_string();
-
-        let key_schema = parse_key_schema(&body["KeySchema"])?;
-        let attribute_definitions = parse_attribute_definitions(&body["AttributeDefinitions"])?;
-
-        // Validate that key schema attributes are defined
-        for ks in &key_schema {
-            if !attribute_definitions
-                .iter()
-                .any(|ad| ad.attribute_name == ks.attribute_name)
-            {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!(
-                        "One or more parameter values were invalid: \
-                         Some index key attributes are not defined in AttributeDefinitions. \
-                         Keys: [{}], AttributeDefinitions: [{}]",
-                        ks.attribute_name,
-                        attribute_definitions
-                            .iter()
-                            .map(|ad| ad.attribute_name.as_str())
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ),
-                ));
-            }
-        }
-
-        let billing_mode = body["BillingMode"]
-            .as_str()
-            .unwrap_or("PROVISIONED")
-            .to_string();
-
-        let provisioned_throughput = if billing_mode == "PAY_PER_REQUEST" {
-            ProvisionedThroughput {
-                read_capacity_units: 0,
-                write_capacity_units: 0,
-            }
-        } else {
-            parse_provisioned_throughput(&body["ProvisionedThroughput"])?
-        };
-
-        let gsi = parse_gsi(&body["GlobalSecondaryIndexes"]);
-        let lsi = parse_lsi(&body["LocalSecondaryIndexes"]);
-        let tags = parse_tags(&body["Tags"]);
-
-        // Parse StreamSpecification
-        let (stream_enabled, stream_view_type) =
-            if let Some(stream_spec) = body.get("StreamSpecification") {
-                let enabled = stream_spec
-                    .get("StreamEnabled")
-                    .and_then(|v| v.as_bool())
-                    .unwrap_or(false);
-                let view_type = if enabled {
-                    stream_spec
-                        .get("StreamViewType")
-                        .and_then(|v| v.as_str())
-                        .map(|s| s.to_string())
-                } else {
-                    None
-                };
-                (enabled, view_type)
-            } else {
-                (false, None)
-            };
-
-        // Parse SSESpecification
-        let (sse_type, sse_kms_key_arn) = if let Some(sse_spec) = body.get("SSESpecification") {
-            let enabled = sse_spec
-                .get("Enabled")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(false);
-            if enabled {
-                let sse_type = sse_spec
-                    .get("SSEType")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("KMS")
-                    .to_string();
-                let kms_key = sse_spec
-                    .get("KMSMasterKeyId")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string());
-                (Some(sse_type), kms_key)
-            } else {
-                (None, None)
-            }
-        } else {
-            (None, None)
-        };
-
-        let mut state = self.state.write();
-
-        if state.tables.contains_key(&table_name) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ResourceInUseException",
-                format!("Table already exists: {table_name}"),
-            ));
-        }
-
-        let now = Utc::now();
-        let arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}",
-            state.region, state.account_id, table_name
-        );
-        let stream_arn = if stream_enabled {
-            Some(format!(
-                "arn:aws:dynamodb:{}:{}:table/{}/stream/{}",
-                state.region,
-                state.account_id,
-                table_name,
-                now.format("%Y-%m-%dT%H:%M:%S.%3f")
-            ))
-        } else {
-            None
-        };
-
-        let table = DynamoTable {
-            name: table_name.clone(),
-            arn: arn.clone(),
-            key_schema: key_schema.clone(),
-            attribute_definitions: attribute_definitions.clone(),
-            provisioned_throughput: provisioned_throughput.clone(),
-            items: Vec::new(),
-            gsi: gsi.clone(),
-            lsi: lsi.clone(),
-            tags,
-            created_at: now,
-            status: "ACTIVE".to_string(),
-            item_count: 0,
-            size_bytes: 0,
-            billing_mode: billing_mode.clone(),
-            ttl_attribute: None,
-            ttl_enabled: false,
-            resource_policy: None,
-            pitr_enabled: false,
-            kinesis_destinations: Vec::new(),
-            contributor_insights_status: "DISABLED".to_string(),
-            contributor_insights_counters: HashMap::new(),
-            stream_enabled,
-            stream_view_type,
-            stream_arn,
-            stream_records: Arc::new(RwLock::new(Vec::new())),
-            sse_type,
-            sse_kms_key_arn,
-        };
-
-        state.tables.insert(table_name, table);
-
-        let table_desc = build_table_description_json(
-            &arn,
-            &key_schema,
-            &attribute_definitions,
-            &provisioned_throughput,
-            &gsi,
-            &lsi,
-            &billing_mode,
-            now,
-            0,
-            0,
-            "ACTIVE",
-        );
-
-        Self::ok_json(json!({ "TableDescription": table_desc }))
-    }
-
-    fn delete_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let mut state = self.state.write();
-        let table = state.tables.remove(table_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ResourceNotFoundException",
-                format!("Requested resource not found: Table: {table_name} not found"),
-            )
-        })?;
-
-        let table_desc = build_table_description_json(
-            &table.arn,
-            &table.key_schema,
-            &table.attribute_definitions,
-            &table.provisioned_throughput,
-            &table.gsi,
-            &table.lsi,
-            &table.billing_mode,
-            table.created_at,
-            table.item_count,
-            table.size_bytes,
-            "DELETING",
-        );
-
-        Self::ok_json(json!({ "TableDescription": table_desc }))
-    }
-
-    fn describe_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let table_desc = build_table_description(table);
-
-        Self::ok_json(json!({ "Table": table_desc }))
-    }
-
-    fn list_tables(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-
-        validate_optional_string_length(
-            "exclusiveStartTableName",
-            body["ExclusiveStartTableName"].as_str(),
-            3,
-            255,
-        )?;
-        validate_optional_range_i64("limit", body["Limit"].as_i64(), 1, 100)?;
-
-        let limit = body["Limit"].as_i64().unwrap_or(100) as usize;
-        let exclusive_start = body["ExclusiveStartTableName"]
-            .as_str()
-            .map(|s| s.to_string());
-
-        let state = self.state.read();
-        let mut names: Vec<&String> = state.tables.keys().collect();
-        names.sort();
-
-        let start_idx = match &exclusive_start {
-            Some(start) => names
-                .iter()
-                .position(|n| n.as_str() > start.as_str())
-                .unwrap_or(names.len()),
-            None => 0,
-        };
-
-        let page: Vec<&str> = names
-            .iter()
-            .skip(start_idx)
-            .take(limit)
-            .map(|n| n.as_str())
-            .collect();
-
-        let mut result = json!({ "TableNames": page });
-
-        if start_idx + limit < names.len() {
-            if let Some(last) = page.last() {
-                result["LastEvaluatedTableName"] = json!(last);
-            }
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn update_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let mut state = self.state.write();
-        let table = state.tables.get_mut(table_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ResourceNotFoundException",
-                format!("Requested resource not found: Table: {table_name} not found"),
-            )
-        })?;
-
-        if let Some(pt) = body.get("ProvisionedThroughput") {
-            if let Ok(throughput) = parse_provisioned_throughput(pt) {
-                table.provisioned_throughput = throughput;
-            }
-        }
-
-        if let Some(bm) = body["BillingMode"].as_str() {
-            table.billing_mode = bm.to_string();
-        }
-
-        // Handle SSESpecification update
-        if let Some(sse_spec) = body.get("SSESpecification") {
-            let enabled = sse_spec
-                .get("Enabled")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(false);
-            if enabled {
-                table.sse_type = Some(
-                    sse_spec
-                        .get("SSEType")
-                        .and_then(|v| v.as_str())
-                        .unwrap_or("KMS")
-                        .to_string(),
-                );
-                table.sse_kms_key_arn = sse_spec
-                    .get("KMSMasterKeyId")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string());
-            } else {
-                table.sse_type = None;
-                table.sse_kms_key_arn = None;
-            }
-        }
-
-        let table_desc = build_table_description_json(
-            &table.arn,
-            &table.key_schema,
-            &table.attribute_definitions,
-            &table.provisioned_throughput,
-            &table.gsi,
-            &table.lsi,
-            &table.billing_mode,
-            table.created_at,
-            table.item_count,
-            table.size_bytes,
-            &table.status,
-        );
-
-        Self::ok_json(json!({ "TableDescription": table_desc }))
-    }
-
-    // ── Item Operations ─────────────────────────────────────────────────
-
-    fn put_item(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        // --- Parse request body and expression attributes WITHOUT holding any lock ---
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let item = require_object(&body, "Item")?;
-        let condition = body["ConditionExpression"].as_str().map(|s| s.to_string());
-        let expr_attr_names = parse_expression_attribute_names(&body);
-        let expr_attr_values = parse_expression_attribute_values(&body);
-        let return_values = body["ReturnValues"].as_str().unwrap_or("NONE").to_string();
-
-        // --- Acquire write lock ONLY for validation + mutation ---
-        // Capture kinesis delivery info alongside the return value
-        let (old_item, kinesis_info) = {
-            let mut state = self.state.write();
-            let region = state.region.clone();
-            let table = get_table_mut(&mut state.tables, table_name)?;
-
-            validate_key_in_item(table, &item)?;
-
-            let key = extract_key(table, &item);
-            let existing_idx = table.find_item_index(&key);
-
-            if let Some(ref cond) = condition {
-                let existing = existing_idx.map(|i| &table.items[i]);
-                evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values)?;
-            }
-
-            let old_item_for_return = if return_values == "ALL_OLD" {
-                existing_idx.map(|i| table.items[i].clone())
-            } else {
-                None
-            };
-
-            // Capture old item for stream/kinesis if needed
-            let needs_change_capture = table.stream_enabled
-                || table
-                    .kinesis_destinations
-                    .iter()
-                    .any(|d| d.destination_status == "ACTIVE");
-            let old_item_for_stream = if needs_change_capture {
-                existing_idx.map(|i| table.items[i].clone())
-            } else {
-                None
-            };
-
-            let is_modify = existing_idx.is_some();
-
-            if let Some(idx) = existing_idx {
-                table.items[idx] = item.clone();
-            } else {
-                table.items.push(item.clone());
-            }
-
-            table.record_item_access(&item);
-            table.recalculate_stats();
-
-            let event_name = if is_modify { "MODIFY" } else { "INSERT" };
-            let key = extract_key(table, &item);
-
-            // Generate stream record
-            if table.stream_enabled {
-                if let Some(record) = crate::streams::generate_stream_record(
-                    table,
-                    event_name,
-                    key.clone(),
-                    old_item_for_stream.clone(),
-                    Some(item.clone()),
-                    &region,
-                ) {
-                    crate::streams::add_stream_record(table, record);
-                }
-            }
-
-            // Capture kinesis delivery info (delivered after lock release)
-            let kinesis_info = if table
-                .kinesis_destinations
-                .iter()
-                .any(|d| d.destination_status == "ACTIVE")
-            {
-                Some((
-                    table.clone(),
-                    event_name.to_string(),
-                    key,
-                    old_item_for_stream,
-                    Some(item.clone()),
-                ))
-            } else {
-                None
-            };
-
-            (old_item_for_return, kinesis_info)
-        };
-        // --- Write lock released, build response ---
-
-        // Deliver to Kinesis destinations outside the lock
-        if let Some((table, event_name, keys, old_image, new_image)) = kinesis_info {
-            self.deliver_to_kinesis_destinations(
-                &table,
-                &event_name,
-                &keys,
-                old_image.as_ref(),
-                new_image.as_ref(),
-            );
-        }
-
-        let mut result = json!({});
-        if let Some(old) = old_item {
-            result["Attributes"] = json!(old);
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn get_item(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        // --- Parse request body WITHOUT holding any lock ---
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let key = require_object(&body, "Key")?;
-
-        // --- Use a read lock for the lookup (allows concurrent GetItem calls) ---
-        let (result, needs_insights) = {
-            let state = self.state.read();
-            let table = get_table(&state.tables, table_name)?;
-            let needs_insights = table.contributor_insights_status == "ENABLED";
-
-            let result = match table.find_item_index(&key) {
-                Some(idx) => {
-                    let item = &table.items[idx];
-                    let projected = project_item(item, &body);
-                    json!({ "Item": projected })
-                }
-                None => json!({}),
-            };
-            (result, needs_insights)
-        };
-        // --- Read lock released ---
-
-        // Only acquire write lock if contributor insights tracking is enabled
-        if needs_insights {
-            let mut state = self.state.write();
-            if let Some(table) = state.tables.get_mut(table_name) {
-                table.record_key_access(&key);
-            }
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn delete_item(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-
-        validate_optional_enum_value(
-            "conditionalOperator",
-            &body["ConditionalOperator"],
-            &["AND", "OR"],
-        )?;
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-        validate_optional_enum_value(
-            "returnValues",
-            &body["ReturnValues"],
-            &["NONE", "ALL_OLD", "UPDATED_OLD", "ALL_NEW", "UPDATED_NEW"],
-        )?;
-        validate_optional_enum_value(
-            "returnItemCollectionMetrics",
-            &body["ReturnItemCollectionMetrics"],
-            &["SIZE", "NONE"],
-        )?;
-        validate_optional_enum_value(
-            "returnValuesOnConditionCheckFailure",
-            &body["ReturnValuesOnConditionCheckFailure"],
-            &["ALL_OLD", "NONE"],
-        )?;
-
-        let table_name = require_str(&body, "TableName")?;
-        let key = require_object(&body, "Key")?;
-
-        let (result, kinesis_info) = {
-            let mut state = self.state.write();
-            let region = state.region.clone();
-            let table = get_table_mut(&mut state.tables, table_name)?;
-
-            let condition = body["ConditionExpression"].as_str();
-            let expr_attr_names = parse_expression_attribute_names(&body);
-            let expr_attr_values = parse_expression_attribute_values(&body);
-
-            let existing_idx = table.find_item_index(&key);
-
-            if let Some(cond) = condition {
-                let existing = existing_idx.map(|i| &table.items[i]);
-                evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values)?;
-            }
-
-            let return_values = body["ReturnValues"].as_str().unwrap_or("NONE");
-
-            let mut result = json!({});
-            let mut kinesis_info = None;
-
-            if let Some(idx) = existing_idx {
-                let old_item = table.items[idx].clone();
-                if return_values == "ALL_OLD" {
-                    result["Attributes"] = json!(old_item.clone());
-                }
-
-                // Generate stream record before removing
-                if table.stream_enabled {
-                    if let Some(record) = crate::streams::generate_stream_record(
-                        table,
-                        "REMOVE",
-                        key.clone(),
-                        Some(old_item.clone()),
-                        None,
-                        &region,
-                    ) {
-                        crate::streams::add_stream_record(table, record);
-                    }
-                }
-
-                // Capture kinesis delivery info
-                if table
-                    .kinesis_destinations
-                    .iter()
-                    .any(|d| d.destination_status == "ACTIVE")
-                {
-                    kinesis_info = Some((table.clone(), key.clone(), Some(old_item)));
-                }
-
-                table.items.remove(idx);
-                table.recalculate_stats();
-            }
-
-            let return_consumed = body["ReturnConsumedCapacity"].as_str().unwrap_or("NONE");
-            let return_icm = body["ReturnItemCollectionMetrics"]
-                .as_str()
-                .unwrap_or("NONE");
-
-            if return_consumed == "TOTAL" || return_consumed == "INDEXES" {
-                result["ConsumedCapacity"] = json!({
-                    "TableName": table_name,
-                    "CapacityUnits": 1.0,
-                });
-            }
-
-            if return_icm == "SIZE" {
-                result["ItemCollectionMetrics"] = json!({});
-            }
-
-            (result, kinesis_info)
-        };
-
-        // Deliver to Kinesis destinations outside the lock
-        if let Some((table, keys, old_image)) = kinesis_info {
-            self.deliver_to_kinesis_destinations(&table, "REMOVE", &keys, old_image.as_ref(), None);
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn update_item(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let key = require_object(&body, "Key")?;
-
-        let mut state = self.state.write();
-        let region = state.region.clone();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-
-        validate_key_attributes_in_key(table, &key)?;
-
-        let condition = body["ConditionExpression"].as_str();
-        let expr_attr_names = parse_expression_attribute_names(&body);
-        let expr_attr_values = parse_expression_attribute_values(&body);
-        let update_expression = body["UpdateExpression"].as_str();
-
-        let existing_idx = table.find_item_index(&key);
-
-        if let Some(cond) = condition {
-            let existing = existing_idx.map(|i| &table.items[i]);
-            evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values)?;
-        }
-
-        let return_values = body["ReturnValues"].as_str().unwrap_or("NONE");
-
-        let is_insert = existing_idx.is_none();
-        let idx = match existing_idx {
-            Some(i) => i,
-            None => {
-                let mut new_item = HashMap::new();
-                for (k, v) in &key {
-                    new_item.insert(k.clone(), v.clone());
-                }
-                table.items.push(new_item);
-                table.items.len() - 1
-            }
-        };
-
-        // Capture old item for stream/kinesis (before update)
-        let needs_change_capture = table.stream_enabled
-            || table
-                .kinesis_destinations
-                .iter()
-                .any(|d| d.destination_status == "ACTIVE");
-        let old_item_for_stream = if needs_change_capture {
-            Some(table.items[idx].clone())
-        } else {
-            None
-        };
-
-        let old_item = if return_values == "ALL_OLD" {
-            Some(table.items[idx].clone())
-        } else {
-            None
-        };
-
-        if let Some(expr) = update_expression {
-            apply_update_expression(
-                &mut table.items[idx],
-                expr,
-                &expr_attr_names,
-                &expr_attr_values,
-            )?;
-        }
-
-        let new_item = if return_values == "ALL_NEW" || return_values == "UPDATED_NEW" {
-            Some(table.items[idx].clone())
-        } else {
-            None
-        };
-
-        let event_name = if is_insert { "INSERT" } else { "MODIFY" };
-        let new_item_for_stream = table.items[idx].clone();
-
-        // Generate stream record after update
-        if table.stream_enabled {
-            if let Some(record) = crate::streams::generate_stream_record(
-                table,
-                event_name,
-                key.clone(),
-                old_item_for_stream.clone(),
-                Some(new_item_for_stream.clone()),
-                &region,
-            ) {
-                crate::streams::add_stream_record(table, record);
-            }
-        }
-
-        // Capture kinesis delivery info
-        let kinesis_info = if table
-            .kinesis_destinations
-            .iter()
-            .any(|d| d.destination_status == "ACTIVE")
-        {
-            Some((
-                table.clone(),
-                event_name.to_string(),
-                key.clone(),
-                old_item_for_stream,
-                Some(new_item_for_stream),
-            ))
-        } else {
-            None
-        };
-
-        table.recalculate_stats();
-
-        // Release the write lock (drop `state`)
-        drop(state);
-
-        // Deliver to Kinesis destinations outside the lock
-        if let Some((tbl, ev, keys, old_image, new_image)) = kinesis_info {
-            self.deliver_to_kinesis_destinations(
-                &tbl,
-                &ev,
-                &keys,
-                old_image.as_ref(),
-                new_image.as_ref(),
-            );
-        }
-
-        let mut result = json!({});
-        if let Some(old) = old_item {
-            result["Attributes"] = json!(old);
-        } else if let Some(new) = new_item {
-            result["Attributes"] = json!(new);
-        }
-
-        Self::ok_json(result)
-    }
-
-    // ── Query & Scan ────────────────────────────────────────────────────
-
-    fn query(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let expr_attr_names = parse_expression_attribute_names(&body);
-        let expr_attr_values = parse_expression_attribute_values(&body);
-
-        let key_condition = body["KeyConditionExpression"].as_str();
-        let filter_expression = body["FilterExpression"].as_str();
-        let scan_forward = body["ScanIndexForward"].as_bool().unwrap_or(true);
-        let limit = body["Limit"].as_i64().map(|l| l as usize);
-        let index_name = body["IndexName"].as_str();
-        let exclusive_start_key: Option<HashMap<String, AttributeValue>> =
-            parse_key_map(&body["ExclusiveStartKey"]);
-
-        let (items_to_scan, hash_key_name, range_key_name): (
-            &[HashMap<String, AttributeValue>],
-            String,
-            Option<String>,
-        ) = if let Some(idx_name) = index_name {
-            if let Some(gsi) = table.gsi.iter().find(|g| g.index_name == idx_name) {
-                let hk = gsi
-                    .key_schema
-                    .iter()
-                    .find(|k| k.key_type == "HASH")
-                    .map(|k| k.attribute_name.clone())
-                    .unwrap_or_default();
-                let rk = gsi
-                    .key_schema
-                    .iter()
-                    .find(|k| k.key_type == "RANGE")
-                    .map(|k| k.attribute_name.clone());
-                (&table.items, hk, rk)
-            } else if let Some(lsi) = table.lsi.iter().find(|l| l.index_name == idx_name) {
-                let hk = lsi
-                    .key_schema
-                    .iter()
-                    .find(|k| k.key_type == "HASH")
-                    .map(|k| k.attribute_name.clone())
-                    .unwrap_or_default();
-                let rk = lsi
-                    .key_schema
-                    .iter()
-                    .find(|k| k.key_type == "RANGE")
-                    .map(|k| k.attribute_name.clone());
-                (&table.items, hk, rk)
-            } else {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("The table does not have the specified index: {idx_name}"),
-                ));
-            }
-        } else {
-            (
-                &table.items[..],
-                table.hash_key_name().to_string(),
-                table.range_key_name().map(|s| s.to_string()),
-            )
-        };
-
-        let mut matched: Vec<&HashMap<String, AttributeValue>> = items_to_scan
-            .iter()
-            .filter(|item| {
-                if let Some(kc) = key_condition {
-                    evaluate_key_condition(
-                        kc,
-                        item,
-                        &hash_key_name,
-                        range_key_name.as_deref(),
-                        &expr_attr_names,
-                        &expr_attr_values,
-                    )
-                } else {
-                    true
-                }
-            })
-            .collect();
-
-        if let Some(ref rk) = range_key_name {
-            matched.sort_by(|a, b| {
-                let av = a.get(rk.as_str());
-                let bv = b.get(rk.as_str());
-                compare_attribute_values(av, bv)
-            });
-            if !scan_forward {
-                matched.reverse();
-            }
-        }
-
-        // For GSI queries, we need the table's primary key attributes to uniquely
-        // identify items (GSI keys are not unique).
-        let table_pk_hash = table.hash_key_name().to_string();
-        let table_pk_range = table.range_key_name().map(|s| s.to_string());
-        let is_gsi_query = index_name.is_some()
-            && (hash_key_name != table_pk_hash
-                || range_key_name.as_deref() != table_pk_range.as_deref());
-
-        // Apply ExclusiveStartKey: skip items up to and including the start key.
-        // For GSI queries the start key contains both index keys and table PK, so
-        // we must match on ALL of them to find the exact item.
-        if let Some(ref start_key) = exclusive_start_key {
-            if let Some(pos) = matched.iter().position(|item| {
-                let index_match =
-                    item_matches_key(item, start_key, &hash_key_name, range_key_name.as_deref());
-                if is_gsi_query {
-                    index_match
-                        && item_matches_key(
-                            item,
-                            start_key,
-                            &table_pk_hash,
-                            table_pk_range.as_deref(),
-                        )
-                } else {
-                    index_match
-                }
-            }) {
-                matched = matched.split_off(pos + 1);
-            }
-        }
-
-        if let Some(filter) = filter_expression {
-            matched.retain(|item| {
-                evaluate_filter_expression(filter, item, &expr_attr_names, &expr_attr_values)
-            });
-        }
-
-        let scanned_count = matched.len();
-
-        let has_more = if let Some(lim) = limit {
-            let more = matched.len() > lim;
-            matched.truncate(lim);
-            more
-        } else {
-            false
-        };
-
-        // Build LastEvaluatedKey from the last returned item if there are more results.
-        // For GSI queries, include both the index keys and the table's primary key
-        // so the item can be uniquely identified on resume.
-        let last_evaluated_key = if has_more {
-            matched.last().map(|item| {
-                let mut key =
-                    extract_key_for_schema(item, &hash_key_name, range_key_name.as_deref());
-                if is_gsi_query {
-                    let table_key =
-                        extract_key_for_schema(item, &table_pk_hash, table_pk_range.as_deref());
-                    key.extend(table_key);
-                }
-                key
-            })
-        } else {
-            None
-        };
-
-        // Collect partition key values for contributor insights
-        let insights_enabled = table.contributor_insights_status == "ENABLED";
-        let pk_name = table.hash_key_name().to_string();
-        let accessed_keys: Vec<String> = if insights_enabled {
-            matched
-                .iter()
-                .filter_map(|item| item.get(&pk_name).map(|v| v.to_string()))
-                .collect()
-        } else {
-            Vec::new()
-        };
-
-        let items: Vec<Value> = matched
-            .iter()
-            .map(|item| {
-                let projected = project_item(item, &body);
-                json!(projected)
-            })
-            .collect();
-
-        let mut result = json!({
-            "Items": items,
-            "Count": items.len(),
-            "ScannedCount": scanned_count,
-        });
-
-        if let Some(lek) = last_evaluated_key {
-            result["LastEvaluatedKey"] = json!(lek);
-        }
-
-        drop(state);
-
-        if !accessed_keys.is_empty() {
-            let mut state = self.state.write();
-            if let Some(table) = state.tables.get_mut(table_name) {
-                // Re-check insights status after acquiring write lock in case it
-                // was disabled between the read and write lock acquisitions.
-                if table.contributor_insights_status == "ENABLED" {
-                    for key_str in accessed_keys {
-                        *table
-                            .contributor_insights_counters
-                            .entry(key_str)
-                            .or_insert(0) += 1;
-                    }
-                }
-            }
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn scan(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let expr_attr_names = parse_expression_attribute_names(&body);
-        let expr_attr_values = parse_expression_attribute_values(&body);
-        let filter_expression = body["FilterExpression"].as_str();
-        let limit = body["Limit"].as_i64().map(|l| l as usize);
-        let exclusive_start_key: Option<HashMap<String, AttributeValue>> =
-            parse_key_map(&body["ExclusiveStartKey"]);
-
-        let hash_key_name = table.hash_key_name().to_string();
-        let range_key_name = table.range_key_name().map(|s| s.to_string());
-
-        let mut matched: Vec<&HashMap<String, AttributeValue>> = table.items.iter().collect();
-
-        // Apply ExclusiveStartKey: skip items up to and including the start key
-        if let Some(ref start_key) = exclusive_start_key {
-            if let Some(pos) = matched.iter().position(|item| {
-                item_matches_key(item, start_key, &hash_key_name, range_key_name.as_deref())
-            }) {
-                matched = matched.split_off(pos + 1);
-            }
-        }
-
-        let scanned_count = matched.len();
-
-        if let Some(filter) = filter_expression {
-            matched.retain(|item| {
-                evaluate_filter_expression(filter, item, &expr_attr_names, &expr_attr_values)
-            });
-        }
-
-        let has_more = if let Some(lim) = limit {
-            let more = matched.len() > lim;
-            matched.truncate(lim);
-            more
-        } else {
-            false
-        };
-
-        // Build LastEvaluatedKey from the last returned item if there are more results
-        let last_evaluated_key = if has_more {
-            matched
-                .last()
-                .map(|item| extract_key_for_schema(item, &hash_key_name, range_key_name.as_deref()))
-        } else {
-            None
-        };
-
-        // Collect partition key values for contributor insights
-        let insights_enabled = table.contributor_insights_status == "ENABLED";
-        let pk_name = table.hash_key_name().to_string();
-        let accessed_keys: Vec<String> = if insights_enabled {
-            matched
-                .iter()
-                .filter_map(|item| item.get(&pk_name).map(|v| v.to_string()))
-                .collect()
-        } else {
-            Vec::new()
-        };
-
-        let items: Vec<Value> = matched
-            .iter()
-            .map(|item| {
-                let projected = project_item(item, &body);
-                json!(projected)
-            })
-            .collect();
-
-        let mut result = json!({
-            "Items": items,
-            "Count": items.len(),
-            "ScannedCount": scanned_count,
-        });
-
-        if let Some(lek) = last_evaluated_key {
-            result["LastEvaluatedKey"] = json!(lek);
-        }
-
-        drop(state);
-
-        if !accessed_keys.is_empty() {
-            let mut state = self.state.write();
-            if let Some(table) = state.tables.get_mut(table_name) {
-                // Re-check insights status after acquiring write lock in case it
-                // was disabled between the read and write lock acquisitions.
-                if table.contributor_insights_status == "ENABLED" {
-                    for key_str in accessed_keys {
-                        *table
-                            .contributor_insights_counters
-                            .entry(key_str)
-                            .or_insert(0) += 1;
-                    }
-                }
-            }
-        }
-
-        Self::ok_json(result)
-    }
-
-    // ── Batch Operations ────────────────────────────────────────────────
-
-    fn batch_get_item(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-
-        let return_consumed = body["ReturnConsumedCapacity"].as_str().unwrap_or("NONE");
-
-        let request_items = body["RequestItems"]
-            .as_object()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "RequestItems is required",
-                )
-            })?
-            .clone();
-
-        let state = self.state.read();
-        let mut responses: HashMap<String, Vec<Value>> = HashMap::new();
-        let mut consumed_capacity: Vec<Value> = Vec::new();
-
-        for (table_name, params) in &request_items {
-            let table = get_table(&state.tables, table_name)?;
-            let keys = params["Keys"].as_array().ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "Keys is required",
-                )
-            })?;
-
-            let mut items = Vec::new();
-            for key_val in keys {
-                let key: HashMap<String, AttributeValue> =
-                    serde_json::from_value(key_val.clone()).unwrap_or_default();
-                if let Some(idx) = table.find_item_index(&key) {
-                    items.push(json!(table.items[idx]));
-                }
-            }
-            responses.insert(table_name.clone(), items);
-
-            if return_consumed == "TOTAL" || return_consumed == "INDEXES" {
-                consumed_capacity.push(json!({
-                    "TableName": table_name,
-                    "CapacityUnits": 1.0,
-                }));
-            }
-        }
-
-        let mut result = json!({
-            "Responses": responses,
-            "UnprocessedKeys": {},
-        });
-
-        if return_consumed == "TOTAL" || return_consumed == "INDEXES" {
-            result["ConsumedCapacity"] = json!(consumed_capacity);
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn batch_write_item(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-        validate_optional_enum_value(
-            "returnItemCollectionMetrics",
-            &body["ReturnItemCollectionMetrics"],
-            &["SIZE", "NONE"],
-        )?;
-
-        let return_consumed = body["ReturnConsumedCapacity"].as_str().unwrap_or("NONE");
-        let return_icm = body["ReturnItemCollectionMetrics"]
-            .as_str()
-            .unwrap_or("NONE");
-
-        let request_items = body["RequestItems"]
-            .as_object()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "RequestItems is required",
-                )
-            })?
-            .clone();
-
-        let mut state = self.state.write();
-        let mut consumed_capacity: Vec<Value> = Vec::new();
-        let mut item_collection_metrics: HashMap<String, Vec<Value>> = HashMap::new();
-
-        for (table_name, requests) in &request_items {
-            let table = state.tables.get_mut(table_name.as_str()).ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ResourceNotFoundException",
-                    format!("Requested resource not found: Table: {table_name} not found"),
-                )
-            })?;
-
-            let reqs = requests.as_array().ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "Request list must be an array",
-                )
-            })?;
-
-            for request in reqs {
-                if let Some(put_req) = request.get("PutRequest") {
-                    let item: HashMap<String, AttributeValue> =
-                        serde_json::from_value(put_req["Item"].clone()).unwrap_or_default();
-                    let key = extract_key(table, &item);
-                    if let Some(idx) = table.find_item_index(&key) {
-                        table.items[idx] = item;
-                    } else {
-                        table.items.push(item);
-                    }
-                } else if let Some(del_req) = request.get("DeleteRequest") {
-                    let key: HashMap<String, AttributeValue> =
-                        serde_json::from_value(del_req["Key"].clone()).unwrap_or_default();
-                    if let Some(idx) = table.find_item_index(&key) {
-                        table.items.remove(idx);
-                    }
-                }
-            }
-
-            table.recalculate_stats();
-
-            if return_consumed == "TOTAL" || return_consumed == "INDEXES" {
-                consumed_capacity.push(json!({
-                    "TableName": table_name,
-                    "CapacityUnits": 1.0,
-                }));
-            }
-
-            if return_icm == "SIZE" {
-                item_collection_metrics.insert(table_name.clone(), vec![]);
-            }
-        }
-
-        let mut result = json!({
-            "UnprocessedItems": {},
-        });
-
-        if return_consumed == "TOTAL" || return_consumed == "INDEXES" {
-            result["ConsumedCapacity"] = json!(consumed_capacity);
-        }
-
-        if return_icm == "SIZE" {
-            result["ItemCollectionMetrics"] = json!(item_collection_metrics);
-        }
-
-        Self::ok_json(result)
-    }
-
-    // ── Tags ────────────────────────────────────────────────────────────
-
-    fn tag_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let resource_arn = require_str(&body, "ResourceArn")?;
-        validate_required("Tags", &body["Tags"])?;
-
-        let mut state = self.state.write();
-        let table = find_table_by_arn_mut(&mut state.tables, resource_arn)?;
-
-        fakecloud_core::tags::apply_tags(&mut table.tags, &body, "Tags", "Key", "Value").map_err(
-            |f| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("{f} must be a list"),
-                )
-            },
-        )?;
-
-        Self::ok_json(json!({}))
-    }
-
-    fn untag_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let resource_arn = require_str(&body, "ResourceArn")?;
-        validate_required("TagKeys", &body["TagKeys"])?;
-
-        let mut state = self.state.write();
-        let table = find_table_by_arn_mut(&mut state.tables, resource_arn)?;
-
-        fakecloud_core::tags::remove_tags(&mut table.tags, &body, "TagKeys").map_err(|f| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                format!("{f} must be a list"),
-            )
-        })?;
-
-        Self::ok_json(json!({}))
-    }
-
-    fn list_tags_of_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let resource_arn = require_str(&body, "ResourceArn")?;
-
-        let state = self.state.read();
-        let table = find_table_by_arn(&state.tables, resource_arn)?;
-
-        let tags = fakecloud_core::tags::tags_to_json(&table.tags, "Key", "Value");
-
-        Self::ok_json(json!({ "Tags": tags }))
-    }
-
-    // ── Transactions ────────────────────────────────────────────────────
-
-    fn transact_get_items(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-        let transact_items = body["TransactItems"].as_array().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TransactItems is required",
-            )
-        })?;
-
-        let state = self.state.read();
-        let mut responses: Vec<Value> = Vec::new();
-
-        for ti in transact_items {
-            let get = &ti["Get"];
-            let table_name = get["TableName"].as_str().ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "TableName is required in Get",
-                )
-            })?;
-            let key: HashMap<String, AttributeValue> =
-                serde_json::from_value(get["Key"].clone()).unwrap_or_default();
-
-            let table = get_table(&state.tables, table_name)?;
-            match table.find_item_index(&key) {
-                Some(idx) => {
-                    responses.push(json!({ "Item": table.items[idx] }));
-                }
-                None => {
-                    responses.push(json!({}));
-                }
-            }
-        }
-
-        Self::ok_json(json!({ "Responses": responses }))
-    }
-
-    fn transact_write_items(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length(
-            "clientRequestToken",
-            body["ClientRequestToken"].as_str(),
-            1,
-            36,
-        )?;
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-        validate_optional_enum_value(
-            "returnItemCollectionMetrics",
-            &body["ReturnItemCollectionMetrics"],
-            &["SIZE", "NONE"],
-        )?;
-        let transact_items = body["TransactItems"].as_array().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TransactItems is required",
-            )
-        })?;
-
-        let mut state = self.state.write();
-
-        // First pass: validate all conditions
-        let mut cancellation_reasons: Vec<Value> = Vec::new();
-        let mut any_failed = false;
-
-        for ti in transact_items {
-            if let Some(put) = ti.get("Put") {
-                let table_name = put["TableName"].as_str().unwrap_or_default();
-                let item: HashMap<String, AttributeValue> =
-                    serde_json::from_value(put["Item"].clone()).unwrap_or_default();
-                let condition = put["ConditionExpression"].as_str();
-
-                if let Some(cond) = condition {
-                    let table = get_table(&state.tables, table_name)?;
-                    let expr_attr_names = parse_expression_attribute_names(put);
-                    let expr_attr_values = parse_expression_attribute_values(put);
-                    let key = extract_key(table, &item);
-                    let existing = table.find_item_index(&key).map(|i| &table.items[i]);
-                    if evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values)
-                        .is_err()
-                    {
-                        cancellation_reasons.push(json!({
-                            "Code": "ConditionalCheckFailed",
-                            "Message": "The conditional request failed"
-                        }));
-                        any_failed = true;
-                        continue;
-                    }
-                }
-                cancellation_reasons.push(json!({ "Code": "None" }));
-            } else if let Some(delete) = ti.get("Delete") {
-                let table_name = delete["TableName"].as_str().unwrap_or_default();
-                let key: HashMap<String, AttributeValue> =
-                    serde_json::from_value(delete["Key"].clone()).unwrap_or_default();
-                let condition = delete["ConditionExpression"].as_str();
-
-                if let Some(cond) = condition {
-                    let table = get_table(&state.tables, table_name)?;
-                    let expr_attr_names = parse_expression_attribute_names(delete);
-                    let expr_attr_values = parse_expression_attribute_values(delete);
-                    let existing = table.find_item_index(&key).map(|i| &table.items[i]);
-                    if evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values)
-                        .is_err()
-                    {
-                        cancellation_reasons.push(json!({
-                            "Code": "ConditionalCheckFailed",
-                            "Message": "The conditional request failed"
-                        }));
-                        any_failed = true;
-                        continue;
-                    }
-                }
-                cancellation_reasons.push(json!({ "Code": "None" }));
-            } else if let Some(update) = ti.get("Update") {
-                let table_name = update["TableName"].as_str().unwrap_or_default();
-                let key: HashMap<String, AttributeValue> =
-                    serde_json::from_value(update["Key"].clone()).unwrap_or_default();
-                let condition = update["ConditionExpression"].as_str();
-
-                if let Some(cond) = condition {
-                    let table = get_table(&state.tables, table_name)?;
-                    let expr_attr_names = parse_expression_attribute_names(update);
-                    let expr_attr_values = parse_expression_attribute_values(update);
-                    let existing = table.find_item_index(&key).map(|i| &table.items[i]);
-                    if evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values)
-                        .is_err()
-                    {
-                        cancellation_reasons.push(json!({
-                            "Code": "ConditionalCheckFailed",
-                            "Message": "The conditional request failed"
-                        }));
-                        any_failed = true;
-                        continue;
-                    }
-                }
-                cancellation_reasons.push(json!({ "Code": "None" }));
-            } else if let Some(check) = ti.get("ConditionCheck") {
-                let table_name = check["TableName"].as_str().unwrap_or_default();
-                let key: HashMap<String, AttributeValue> =
-                    serde_json::from_value(check["Key"].clone()).unwrap_or_default();
-                let cond = check["ConditionExpression"].as_str().unwrap_or_default();
-
-                let table = get_table(&state.tables, table_name)?;
-                let expr_attr_names = parse_expression_attribute_names(check);
-                let expr_attr_values = parse_expression_attribute_values(check);
-                let existing = table.find_item_index(&key).map(|i| &table.items[i]);
-                if evaluate_condition(cond, existing, &expr_attr_names, &expr_attr_values).is_err()
-                {
-                    cancellation_reasons.push(json!({
-                        "Code": "ConditionalCheckFailed",
-                        "Message": "The conditional request failed"
-                    }));
-                    any_failed = true;
-                    continue;
-                }
-                cancellation_reasons.push(json!({ "Code": "None" }));
-            } else {
-                cancellation_reasons.push(json!({ "Code": "None" }));
-            }
-        }
-
-        if any_failed {
-            let error_body = json!({
-                "__type": "TransactionCanceledException",
-                "message": "Transaction cancelled, please refer cancellation reasons for specific reasons [ConditionalCheckFailed]",
-                "CancellationReasons": cancellation_reasons
-            });
-            return Ok(AwsResponse::json(
-                StatusCode::BAD_REQUEST,
-                serde_json::to_vec(&error_body).unwrap(),
-            ));
-        }
-
-        // Second pass: apply all writes
-        for ti in transact_items {
-            if let Some(put) = ti.get("Put") {
-                let table_name = put["TableName"].as_str().unwrap_or_default();
-                let item: HashMap<String, AttributeValue> =
-                    serde_json::from_value(put["Item"].clone()).unwrap_or_default();
-                let table = get_table_mut(&mut state.tables, table_name)?;
-                let key = extract_key(table, &item);
-                if let Some(idx) = table.find_item_index(&key) {
-                    table.items[idx] = item;
-                } else {
-                    table.items.push(item);
-                }
-                table.recalculate_stats();
-            } else if let Some(delete) = ti.get("Delete") {
-                let table_name = delete["TableName"].as_str().unwrap_or_default();
-                let key: HashMap<String, AttributeValue> =
-                    serde_json::from_value(delete["Key"].clone()).unwrap_or_default();
-                let table = get_table_mut(&mut state.tables, table_name)?;
-                if let Some(idx) = table.find_item_index(&key) {
-                    table.items.remove(idx);
-                }
-                table.recalculate_stats();
-            } else if let Some(update) = ti.get("Update") {
-                let table_name = update["TableName"].as_str().unwrap_or_default();
-                let key: HashMap<String, AttributeValue> =
-                    serde_json::from_value(update["Key"].clone()).unwrap_or_default();
-                let update_expression = update["UpdateExpression"].as_str();
-                let expr_attr_names = parse_expression_attribute_names(update);
-                let expr_attr_values = parse_expression_attribute_values(update);
-
-                let table = get_table_mut(&mut state.tables, table_name)?;
-                let idx = match table.find_item_index(&key) {
-                    Some(i) => i,
-                    None => {
-                        let mut new_item = HashMap::new();
-                        for (k, v) in &key {
-                            new_item.insert(k.clone(), v.clone());
-                        }
-                        table.items.push(new_item);
-                        table.items.len() - 1
-                    }
-                };
-
-                if let Some(expr) = update_expression {
-                    apply_update_expression(
-                        &mut table.items[idx],
-                        expr,
-                        &expr_attr_names,
-                        &expr_attr_values,
-                    )?;
-                }
-                table.recalculate_stats();
-            }
-            // ConditionCheck: no write needed
-        }
-
-        Self::ok_json(json!({}))
-    }
-
-    fn execute_statement(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let statement = require_str(&body, "Statement")?;
-        let parameters = body["Parameters"].as_array().cloned().unwrap_or_default();
-
-        execute_partiql_statement(&self.state, statement, &parameters)
-    }
-
-    fn batch_execute_statement(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-        let statements = body["Statements"].as_array().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "Statements is required",
-            )
-        })?;
-
-        let mut responses: Vec<Value> = Vec::new();
-        for stmt_obj in statements {
-            let statement = stmt_obj["Statement"].as_str().unwrap_or_default();
-            let parameters = stmt_obj["Parameters"]
-                .as_array()
-                .cloned()
-                .unwrap_or_default();
-
-            match execute_partiql_statement(&self.state, statement, &parameters) {
-                Ok(resp) => {
-                    let resp_body: Value = serde_json::from_slice(&resp.body).unwrap_or_default();
-                    responses.push(resp_body);
-                }
-                Err(e) => {
-                    responses.push(json!({
-                        "Error": {
-                            "Code": "ValidationException",
-                            "Message": e.to_string()
-                        }
-                    }));
-                }
-            }
-        }
-
-        Self::ok_json(json!({ "Responses": responses }))
-    }
-
-    fn execute_transaction(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length(
-            "clientRequestToken",
-            body["ClientRequestToken"].as_str(),
-            1,
-            36,
-        )?;
-        validate_optional_enum_value(
-            "returnConsumedCapacity",
-            &body["ReturnConsumedCapacity"],
-            &["INDEXES", "TOTAL", "NONE"],
-        )?;
-        let transact_statements = body["TransactStatements"].as_array().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TransactStatements is required",
-            )
-        })?;
-
-        // Collect all results; if any fail, return TransactionCanceledException
-        let mut results: Vec<Result<Value, String>> = Vec::new();
-        for stmt_obj in transact_statements {
-            let statement = stmt_obj["Statement"].as_str().unwrap_or_default();
-            let parameters = stmt_obj["Parameters"]
-                .as_array()
-                .cloned()
-                .unwrap_or_default();
-
-            match execute_partiql_statement(&self.state, statement, &parameters) {
-                Ok(resp) => {
-                    let resp_body: Value = serde_json::from_slice(&resp.body).unwrap_or_default();
-                    results.push(Ok(resp_body));
-                }
-                Err(e) => {
-                    results.push(Err(e.to_string()));
-                }
-            }
-        }
-
-        let any_failed = results.iter().any(|r| r.is_err());
-        if any_failed {
-            let reasons: Vec<Value> = results
-                .iter()
-                .map(|r| match r {
-                    Ok(_) => json!({ "Code": "None" }),
-                    Err(msg) => json!({
-                        "Code": "ValidationException",
-                        "Message": msg
-                    }),
-                })
-                .collect();
-            let error_body = json!({
-                "__type": "TransactionCanceledException",
-                "message": "Transaction cancelled due to validation errors",
-                "CancellationReasons": reasons
-            });
-            return Ok(AwsResponse::json(
-                StatusCode::BAD_REQUEST,
-                serde_json::to_vec(&error_body).unwrap(),
-            ));
-        }
-
-        let responses: Vec<Value> = results.into_iter().filter_map(|r| r.ok()).collect();
-        Self::ok_json(json!({ "Responses": responses }))
-    }
-
-    // ── TTL ─────────────────────────────────────────────────────────────
-
-    fn update_time_to_live(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let spec = &body["TimeToLiveSpecification"];
-        let attr_name = spec["AttributeName"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TimeToLiveSpecification.AttributeName is required",
-            )
-        })?;
-        let enabled = spec["Enabled"].as_bool().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TimeToLiveSpecification.Enabled is required",
-            )
-        })?;
-
-        let mut state = self.state.write();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-
-        if enabled {
-            table.ttl_attribute = Some(attr_name.to_string());
-            table.ttl_enabled = true;
-        } else {
-            table.ttl_enabled = false;
-        }
-
-        Self::ok_json(json!({
-            "TimeToLiveSpecification": {
-                "AttributeName": attr_name,
-                "Enabled": enabled
-            }
-        }))
-    }
-
-    fn describe_time_to_live(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let status = if table.ttl_enabled {
-            "ENABLED"
-        } else {
-            "DISABLED"
-        };
-
-        let mut desc = json!({
-            "TimeToLiveDescription": {
-                "TimeToLiveStatus": status
-            }
-        });
-
-        if let Some(ref attr) = table.ttl_attribute {
-            desc["TimeToLiveDescription"]["AttributeName"] = json!(attr);
-        }
-
-        Self::ok_json(desc)
-    }
-
-    // ── Resource Policies ───────────────────────────────────────────────
-
-    fn put_resource_policy(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let resource_arn = require_str(&body, "ResourceArn")?;
-        let policy = require_str(&body, "Policy")?;
-
-        let mut state = self.state.write();
-        let table = find_table_by_arn_mut(&mut state.tables, resource_arn)?;
-        table.resource_policy = Some(policy.to_string());
-
-        let revision_id = uuid::Uuid::new_v4().to_string();
-        Self::ok_json(json!({ "RevisionId": revision_id }))
-    }
-
-    fn get_resource_policy(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let resource_arn = require_str(&body, "ResourceArn")?;
-
-        let state = self.state.read();
-        let table = find_table_by_arn(&state.tables, resource_arn)?;
-
-        match &table.resource_policy {
-            Some(policy) => {
-                let revision_id = uuid::Uuid::new_v4().to_string();
-                Self::ok_json(json!({
-                    "Policy": policy,
-                    "RevisionId": revision_id
-                }))
-            }
-            None => Self::ok_json(json!({ "Policy": null })),
-        }
-    }
-
-    fn delete_resource_policy(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let resource_arn = require_str(&body, "ResourceArn")?;
-
-        let mut state = self.state.write();
-        let table = find_table_by_arn_mut(&mut state.tables, resource_arn)?;
-        table.resource_policy = None;
-
-        Self::ok_json(json!({}))
-    }
-
-    // ── Stubs ──────────────────────────────────────────────────────────
-
-    fn describe_endpoints(&self, _req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        Self::ok_json(json!({
-            "Endpoints": [{
-                "Address": "dynamodb.us-east-1.amazonaws.com",
-                "CachePeriodInMinutes": 1440
-            }]
-        }))
-    }
-
-    fn describe_limits(&self, _req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        Self::ok_json(json!({
-            "AccountMaxReadCapacityUnits": 80000,
-            "AccountMaxWriteCapacityUnits": 80000,
-            "TableMaxReadCapacityUnits": 40000,
-            "TableMaxWriteCapacityUnits": 40000
-        }))
-    }
-
-    // ── Backups ────────────────────────────────────────────────────────
-
-    fn create_backup(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let backup_name = require_str(&body, "BackupName")?;
-
-        let mut state = self.state.write();
-        let table = get_table(&state.tables, table_name)?;
-
-        let backup_arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}/backup/{}",
-            state.region,
-            state.account_id,
-            table_name,
-            Utc::now().format("%Y%m%d%H%M%S")
-        );
-        let now = Utc::now();
-
-        let backup = BackupDescription {
-            backup_arn: backup_arn.clone(),
-            backup_name: backup_name.to_string(),
-            table_name: table_name.to_string(),
-            table_arn: table.arn.clone(),
-            backup_status: "AVAILABLE".to_string(),
-            backup_type: "USER".to_string(),
-            backup_creation_date: now,
-            key_schema: table.key_schema.clone(),
-            attribute_definitions: table.attribute_definitions.clone(),
-            provisioned_throughput: table.provisioned_throughput.clone(),
-            billing_mode: table.billing_mode.clone(),
-            item_count: table.item_count,
-            size_bytes: table.size_bytes,
-            items: table.items.clone(),
-        };
-
-        state.backups.insert(backup_arn.clone(), backup);
-
-        Self::ok_json(json!({
-            "BackupDetails": {
-                "BackupArn": backup_arn,
-                "BackupName": backup_name,
-                "BackupStatus": "AVAILABLE",
-                "BackupType": "USER",
-                "BackupCreationDateTime": now.timestamp() as f64,
-                "BackupSizeBytes": 0
-            }
-        }))
-    }
-
-    fn delete_backup(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let backup_arn = require_str(&body, "BackupArn")?;
-
-        let mut state = self.state.write();
-        let backup = state.backups.remove(backup_arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "BackupNotFoundException",
-                format!("Backup not found: {backup_arn}"),
-            )
-        })?;
-
-        Self::ok_json(json!({
-            "BackupDescription": {
-                "BackupDetails": {
-                    "BackupArn": backup.backup_arn,
-                    "BackupName": backup.backup_name,
-                    "BackupStatus": "DELETED",
-                    "BackupType": backup.backup_type,
-                    "BackupCreationDateTime": backup.backup_creation_date.timestamp() as f64,
-                    "BackupSizeBytes": backup.size_bytes
-                },
-                "SourceTableDetails": {
-                    "TableName": backup.table_name,
-                    "TableArn": backup.table_arn,
-                    "TableId": uuid::Uuid::new_v4().to_string(),
-                    "KeySchema": backup.key_schema.iter().map(|ks| json!({
-                        "AttributeName": ks.attribute_name,
-                        "KeyType": ks.key_type
-                    })).collect::<Vec<_>>(),
-                    "TableCreationDateTime": backup.backup_creation_date.timestamp() as f64,
-                    "ProvisionedThroughput": {
-                        "ReadCapacityUnits": backup.provisioned_throughput.read_capacity_units,
-                        "WriteCapacityUnits": backup.provisioned_throughput.write_capacity_units
-                    },
-                    "ItemCount": backup.item_count,
-                    "BillingMode": backup.billing_mode,
-                    "TableSizeBytes": backup.size_bytes
-                },
-                "SourceTableFeatureDetails": {}
-            }
-        }))
-    }
-
-    fn describe_backup(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let backup_arn = require_str(&body, "BackupArn")?;
-
-        let state = self.state.read();
-        let backup = state.backups.get(backup_arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "BackupNotFoundException",
-                format!("Backup not found: {backup_arn}"),
-            )
-        })?;
-
-        Self::ok_json(json!({
-            "BackupDescription": {
-                "BackupDetails": {
-                    "BackupArn": backup.backup_arn,
-                    "BackupName": backup.backup_name,
-                    "BackupStatus": backup.backup_status,
-                    "BackupType": backup.backup_type,
-                    "BackupCreationDateTime": backup.backup_creation_date.timestamp() as f64,
-                    "BackupSizeBytes": backup.size_bytes
-                },
-                "SourceTableDetails": {
-                    "TableName": backup.table_name,
-                    "TableArn": backup.table_arn,
-                    "TableId": uuid::Uuid::new_v4().to_string(),
-                    "KeySchema": backup.key_schema.iter().map(|ks| json!({
-                        "AttributeName": ks.attribute_name,
-                        "KeyType": ks.key_type
-                    })).collect::<Vec<_>>(),
-                    "TableCreationDateTime": backup.backup_creation_date.timestamp() as f64,
-                    "ProvisionedThroughput": {
-                        "ReadCapacityUnits": backup.provisioned_throughput.read_capacity_units,
-                        "WriteCapacityUnits": backup.provisioned_throughput.write_capacity_units
-                    },
-                    "ItemCount": backup.item_count,
-                    "BillingMode": backup.billing_mode,
-                    "TableSizeBytes": backup.size_bytes
-                },
-                "SourceTableFeatureDetails": {}
-            }
-        }))
-    }
-
-    fn list_backups(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length("tableName", body["TableName"].as_str(), 1, 1024)?;
-        validate_optional_string_length(
-            "exclusiveStartBackupArn",
-            body["ExclusiveStartBackupArn"].as_str(),
-            37,
-            1024,
-        )?;
-        validate_optional_range_i64("limit", body["Limit"].as_i64(), 1, 100)?;
-        validate_optional_enum_value(
-            "backupType",
-            &body["BackupType"],
-            &["USER", "SYSTEM", "AWS_BACKUP", "ALL"],
-        )?;
-        let table_name = body["TableName"].as_str();
-
-        let state = self.state.read();
-        let summaries: Vec<Value> = state
-            .backups
-            .values()
-            .filter(|b| table_name.is_none() || table_name == Some(b.table_name.as_str()))
-            .map(|b| {
-                json!({
-                    "TableName": b.table_name,
-                    "TableArn": b.table_arn,
-                    "BackupArn": b.backup_arn,
-                    "BackupName": b.backup_name,
-                    "BackupCreationDateTime": b.backup_creation_date.timestamp() as f64,
-                    "BackupStatus": b.backup_status,
-                    "BackupType": b.backup_type,
-                    "BackupSizeBytes": b.size_bytes
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "BackupSummaries": summaries
-        }))
-    }
-
-    fn restore_table_from_backup(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let backup_arn = require_str(&body, "BackupArn")?;
-        let target_table_name = require_str(&body, "TargetTableName")?;
-
-        let mut state = self.state.write();
-        let backup = state.backups.get(backup_arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "BackupNotFoundException",
-                format!("Backup not found: {backup_arn}"),
-            )
-        })?;
-
-        if state.tables.contains_key(target_table_name) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "TableAlreadyExistsException",
-                format!("Table already exists: {target_table_name}"),
-            ));
-        }
-
-        let now = Utc::now();
-        let arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}",
-            state.region, state.account_id, target_table_name
-        );
-
-        let restored_items = backup.items.clone();
-        let mut table = DynamoTable {
-            name: target_table_name.to_string(),
-            arn: arn.clone(),
-            key_schema: backup.key_schema.clone(),
-            attribute_definitions: backup.attribute_definitions.clone(),
-            provisioned_throughput: backup.provisioned_throughput.clone(),
-            items: restored_items,
-            gsi: Vec::new(),
-            lsi: Vec::new(),
-            tags: HashMap::new(),
-            created_at: now,
-            status: "ACTIVE".to_string(),
-            item_count: 0,
-            size_bytes: 0,
-            billing_mode: backup.billing_mode.clone(),
-            ttl_attribute: None,
-            ttl_enabled: false,
-            resource_policy: None,
-            pitr_enabled: false,
-            kinesis_destinations: Vec::new(),
-            contributor_insights_status: "DISABLED".to_string(),
-            contributor_insights_counters: HashMap::new(),
-            stream_enabled: false,
-            stream_view_type: None,
-            stream_arn: None,
-            stream_records: Arc::new(RwLock::new(Vec::new())),
-            sse_type: None,
-            sse_kms_key_arn: None,
-        };
-        table.recalculate_stats();
-
-        let desc = build_table_description(&table);
-        state.tables.insert(target_table_name.to_string(), table);
-
-        Self::ok_json(json!({
-            "TableDescription": desc
-        }))
-    }
-
-    fn restore_table_to_point_in_time(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let target_table_name = require_str(&body, "TargetTableName")?;
-        let source_table_name = body["SourceTableName"].as_str();
-        let source_table_arn = body["SourceTableArn"].as_str();
-
-        let mut state = self.state.write();
-
-        // Resolve source table
-        let source = if let Some(name) = source_table_name {
-            get_table(&state.tables, name)?.clone()
-        } else if let Some(arn) = source_table_arn {
-            find_table_by_arn(&state.tables, arn)?.clone()
-        } else {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "SourceTableName or SourceTableArn is required",
-            ));
-        };
-
-        if state.tables.contains_key(target_table_name) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "TableAlreadyExistsException",
-                format!("Table already exists: {target_table_name}"),
-            ));
-        }
-
-        let now = Utc::now();
-        let arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}",
-            state.region, state.account_id, target_table_name
-        );
-
-        let mut table = DynamoTable {
-            name: target_table_name.to_string(),
-            arn: arn.clone(),
-            key_schema: source.key_schema.clone(),
-            attribute_definitions: source.attribute_definitions.clone(),
-            provisioned_throughput: source.provisioned_throughput.clone(),
-            items: source.items.clone(),
-            gsi: Vec::new(),
-            lsi: Vec::new(),
-            tags: HashMap::new(),
-            created_at: now,
-            status: "ACTIVE".to_string(),
-            item_count: 0,
-            size_bytes: 0,
-            billing_mode: source.billing_mode.clone(),
-            ttl_attribute: None,
-            ttl_enabled: false,
-            resource_policy: None,
-            pitr_enabled: false,
-            kinesis_destinations: Vec::new(),
-            contributor_insights_status: "DISABLED".to_string(),
-            contributor_insights_counters: HashMap::new(),
-            stream_enabled: false,
-            stream_view_type: None,
-            stream_arn: None,
-            stream_records: Arc::new(RwLock::new(Vec::new())),
-            sse_type: None,
-            sse_kms_key_arn: None,
-        };
-        table.recalculate_stats();
-
-        let desc = build_table_description(&table);
-        state.tables.insert(target_table_name.to_string(), table);
-
-        Self::ok_json(json!({
-            "TableDescription": desc
-        }))
-    }
-
-    fn update_continuous_backups(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let pitr_spec = body["PointInTimeRecoverySpecification"]
-            .as_object()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "PointInTimeRecoverySpecification is required",
-                )
-            })?;
-        let enabled = pitr_spec
-            .get("PointInTimeRecoveryEnabled")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        let mut state = self.state.write();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-        table.pitr_enabled = enabled;
-
-        let status = if enabled { "ENABLED" } else { "DISABLED" };
-        Self::ok_json(json!({
-            "ContinuousBackupsDescription": {
-                "ContinuousBackupsStatus": status,
-                "PointInTimeRecoveryDescription": {
-                    "PointInTimeRecoveryStatus": status,
-                    "EarliestRestorableDateTime": Utc::now().timestamp() as f64,
-                    "LatestRestorableDateTime": Utc::now().timestamp() as f64
-                }
-            }
-        }))
-    }
-
-    fn describe_continuous_backups(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let status = if table.pitr_enabled {
-            "ENABLED"
-        } else {
-            "DISABLED"
-        };
-        Self::ok_json(json!({
-            "ContinuousBackupsDescription": {
-                "ContinuousBackupsStatus": status,
-                "PointInTimeRecoveryDescription": {
-                    "PointInTimeRecoveryStatus": status,
-                    "EarliestRestorableDateTime": Utc::now().timestamp() as f64,
-                    "LatestRestorableDateTime": Utc::now().timestamp() as f64
-                }
-            }
-        }))
-    }
-
-    // ── Global Tables ──────────────────────────────────────────────────
-
-    fn create_global_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let global_table_name = require_str(&body, "GlobalTableName")?;
-        validate_string_length("globalTableName", global_table_name, 3, 255)?;
-
-        let replication_group = body["ReplicationGroup"]
-            .as_array()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "ReplicationGroup is required",
-                )
-            })?
-            .iter()
-            .filter_map(|r| {
-                r["RegionName"].as_str().map(|rn| ReplicaDescription {
-                    region_name: rn.to_string(),
-                    replica_status: "ACTIVE".to_string(),
-                })
-            })
-            .collect::<Vec<_>>();
-
-        let mut state = self.state.write();
-
-        if state.global_tables.contains_key(global_table_name) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "GlobalTableAlreadyExistsException",
-                format!("Global table already exists: {global_table_name}"),
-            ));
-        }
-
-        let arn = format!(
-            "arn:aws:dynamodb::{}:global-table/{}",
-            state.account_id, global_table_name
-        );
-        let now = Utc::now();
-
-        let gt = GlobalTableDescription {
-            global_table_name: global_table_name.to_string(),
-            global_table_arn: arn.clone(),
-            global_table_status: "ACTIVE".to_string(),
-            creation_date: now,
-            replication_group: replication_group.clone(),
-        };
-
-        state
-            .global_tables
-            .insert(global_table_name.to_string(), gt);
-
-        Self::ok_json(json!({
-            "GlobalTableDescription": {
-                "GlobalTableName": global_table_name,
-                "GlobalTableArn": arn,
-                "GlobalTableStatus": "ACTIVE",
-                "CreationDateTime": now.timestamp() as f64,
-                "ReplicationGroup": replication_group.iter().map(|r| json!({
-                    "RegionName": r.region_name,
-                    "ReplicaStatus": r.replica_status
-                })).collect::<Vec<_>>()
-            }
-        }))
-    }
-
-    fn describe_global_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let global_table_name = require_str(&body, "GlobalTableName")?;
-        validate_string_length("globalTableName", global_table_name, 3, 255)?;
-
-        let state = self.state.read();
-        let gt = state.global_tables.get(global_table_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "GlobalTableNotFoundException",
-                format!("Global table not found: {global_table_name}"),
-            )
-        })?;
-
-        Self::ok_json(json!({
-            "GlobalTableDescription": {
-                "GlobalTableName": gt.global_table_name,
-                "GlobalTableArn": gt.global_table_arn,
-                "GlobalTableStatus": gt.global_table_status,
-                "CreationDateTime": gt.creation_date.timestamp() as f64,
-                "ReplicationGroup": gt.replication_group.iter().map(|r| json!({
-                    "RegionName": r.region_name,
-                    "ReplicaStatus": r.replica_status
-                })).collect::<Vec<_>>()
-            }
-        }))
-    }
-
-    fn describe_global_table_settings(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let global_table_name = require_str(&body, "GlobalTableName")?;
-        validate_string_length("globalTableName", global_table_name, 3, 255)?;
-
-        let state = self.state.read();
-        let gt = state.global_tables.get(global_table_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "GlobalTableNotFoundException",
-                format!("Global table not found: {global_table_name}"),
-            )
-        })?;
-
-        let replica_settings: Vec<Value> = gt
-            .replication_group
-            .iter()
-            .map(|r| {
-                json!({
-                    "RegionName": r.region_name,
-                    "ReplicaStatus": r.replica_status,
-                    "ReplicaProvisionedReadCapacityUnits": 0,
-                    "ReplicaProvisionedWriteCapacityUnits": 0
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "GlobalTableName": gt.global_table_name,
-            "ReplicaSettings": replica_settings
-        }))
-    }
-
-    fn list_global_tables(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length(
-            "exclusiveStartGlobalTableName",
-            body["ExclusiveStartGlobalTableName"].as_str(),
-            3,
-            255,
-        )?;
-        validate_optional_range_i64("limit", body["Limit"].as_i64(), 1, i64::MAX)?;
-        let limit = body["Limit"].as_i64().unwrap_or(100) as usize;
-
-        let state = self.state.read();
-        let tables: Vec<Value> = state
-            .global_tables
-            .values()
-            .take(limit)
-            .map(|gt| {
-                json!({
-                    "GlobalTableName": gt.global_table_name,
-                    "ReplicationGroup": gt.replication_group.iter().map(|r| json!({
-                        "RegionName": r.region_name
-                    })).collect::<Vec<_>>()
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "GlobalTables": tables
-        }))
-    }
-
-    fn update_global_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let global_table_name = require_str(&body, "GlobalTableName")?;
-        validate_string_length("globalTableName", global_table_name, 3, 255)?;
-        validate_required("replicaUpdates", &body["ReplicaUpdates"])?;
-
-        let mut state = self.state.write();
-        let gt = state
-            .global_tables
-            .get_mut(global_table_name)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "GlobalTableNotFoundException",
-                    format!("Global table not found: {global_table_name}"),
-                )
-            })?;
-
-        if let Some(updates) = body["ReplicaUpdates"].as_array() {
-            for update in updates {
-                if let Some(create) = update["Create"].as_object() {
-                    if let Some(region) = create.get("RegionName").and_then(|v| v.as_str()) {
-                        gt.replication_group.push(ReplicaDescription {
-                            region_name: region.to_string(),
-                            replica_status: "ACTIVE".to_string(),
-                        });
-                    }
-                }
-                if let Some(delete) = update["Delete"].as_object() {
-                    if let Some(region) = delete.get("RegionName").and_then(|v| v.as_str()) {
-                        gt.replication_group.retain(|r| r.region_name != region);
-                    }
-                }
-            }
-        }
-
-        Self::ok_json(json!({
-            "GlobalTableDescription": {
-                "GlobalTableName": gt.global_table_name,
-                "GlobalTableArn": gt.global_table_arn,
-                "GlobalTableStatus": gt.global_table_status,
-                "CreationDateTime": gt.creation_date.timestamp() as f64,
-                "ReplicationGroup": gt.replication_group.iter().map(|r| json!({
-                    "RegionName": r.region_name,
-                    "ReplicaStatus": r.replica_status
-                })).collect::<Vec<_>>()
-            }
-        }))
-    }
-
-    fn update_global_table_settings(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let global_table_name = require_str(&body, "GlobalTableName")?;
-        validate_string_length("globalTableName", global_table_name, 3, 255)?;
-        validate_optional_enum_value(
-            "globalTableBillingMode",
-            &body["GlobalTableBillingMode"],
-            &["PROVISIONED", "PAY_PER_REQUEST"],
-        )?;
-        validate_optional_range_i64(
-            "globalTableProvisionedWriteCapacityUnits",
-            body["GlobalTableProvisionedWriteCapacityUnits"].as_i64(),
-            1,
-            i64::MAX,
-        )?;
-
-        let state = self.state.read();
-        let gt = state.global_tables.get(global_table_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "GlobalTableNotFoundException",
-                format!("Global table not found: {global_table_name}"),
-            )
-        })?;
-
-        let replica_settings: Vec<Value> = gt
-            .replication_group
-            .iter()
-            .map(|r| {
-                json!({
-                    "RegionName": r.region_name,
-                    "ReplicaStatus": r.replica_status,
-                    "ReplicaProvisionedReadCapacityUnits": 0,
-                    "ReplicaProvisionedWriteCapacityUnits": 0
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "GlobalTableName": gt.global_table_name,
-            "ReplicaSettings": replica_settings
-        }))
-    }
-
-    fn describe_table_replica_auto_scaling(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        Self::ok_json(json!({
-            "TableAutoScalingDescription": {
-                "TableName": table.name,
-                "TableStatus": table.status,
-                "Replicas": []
-            }
-        }))
-    }
-
-    fn update_table_replica_auto_scaling(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        Self::ok_json(json!({
-            "TableAutoScalingDescription": {
-                "TableName": table.name,
-                "TableStatus": table.status,
-                "Replicas": []
-            }
-        }))
-    }
-
-    // ── Kinesis Streaming ──────────────────────────────────────────────
-
-    fn enable_kinesis_streaming_destination(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let stream_arn = require_str(&body, "StreamArn")?;
-        let precision = body["EnableKinesisStreamingConfiguration"]
-            ["ApproximateCreationDateTimePrecision"]
-            .as_str()
-            .unwrap_or("MILLISECOND");
-
-        let mut state = self.state.write();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-
-        table.kinesis_destinations.push(KinesisDestination {
-            stream_arn: stream_arn.to_string(),
-            destination_status: "ACTIVE".to_string(),
-            approximate_creation_date_time_precision: precision.to_string(),
-        });
-
-        Self::ok_json(json!({
-            "TableName": table_name,
-            "StreamArn": stream_arn,
-            "DestinationStatus": "ACTIVE",
-            "EnableKinesisStreamingConfiguration": {
-                "ApproximateCreationDateTimePrecision": precision
-            }
-        }))
-    }
-
-    fn disable_kinesis_streaming_destination(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let stream_arn = require_str(&body, "StreamArn")?;
-
-        let mut state = self.state.write();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-
-        if let Some(dest) = table
-            .kinesis_destinations
-            .iter_mut()
-            .find(|d| d.stream_arn == stream_arn)
-        {
-            dest.destination_status = "DISABLED".to_string();
-        }
-
-        Self::ok_json(json!({
-            "TableName": table_name,
-            "StreamArn": stream_arn,
-            "DestinationStatus": "DISABLED"
-        }))
-    }
-
-    fn describe_kinesis_streaming_destination(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let destinations: Vec<Value> = table
-            .kinesis_destinations
-            .iter()
-            .map(|d| {
-                json!({
-                    "StreamArn": d.stream_arn,
-                    "DestinationStatus": d.destination_status,
-                    "ApproximateCreationDateTimePrecision": d.approximate_creation_date_time_precision
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "TableName": table_name,
-            "KinesisDataStreamDestinations": destinations
-        }))
-    }
-
-    fn update_kinesis_streaming_destination(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let stream_arn = require_str(&body, "StreamArn")?;
-        let precision = body["UpdateKinesisStreamingConfiguration"]
-            ["ApproximateCreationDateTimePrecision"]
-            .as_str()
-            .unwrap_or("MILLISECOND");
-
-        let mut state = self.state.write();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-
-        if let Some(dest) = table
-            .kinesis_destinations
-            .iter_mut()
-            .find(|d| d.stream_arn == stream_arn)
-        {
-            dest.approximate_creation_date_time_precision = precision.to_string();
-        }
-
-        Self::ok_json(json!({
-            "TableName": table_name,
-            "StreamArn": stream_arn,
-            "DestinationStatus": "ACTIVE",
-            "UpdateKinesisStreamingConfiguration": {
-                "ApproximateCreationDateTimePrecision": precision
-            }
-        }))
-    }
-
-    // ── Contributor Insights ───────────────────────────────────────────
-
-    fn describe_contributor_insights(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let index_name = body["IndexName"].as_str();
-
-        let state = self.state.read();
-        let table = get_table(&state.tables, table_name)?;
-
-        let top = table.top_contributors(10);
-        let contributors: Vec<Value> = top
-            .iter()
-            .map(|(key, count)| {
-                json!({
-                    "Key": key,
-                    "Count": count
-                })
-            })
-            .collect();
-
-        let mut result = json!({
-            "TableName": table_name,
-            "ContributorInsightsStatus": table.contributor_insights_status,
-            "ContributorInsightsRuleList": ["DynamoDBContributorInsights"],
-            "TopContributors": contributors
-        });
-        if let Some(idx) = index_name {
-            result["IndexName"] = json!(idx);
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn update_contributor_insights(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_name = require_str(&body, "TableName")?;
-        let action = require_str(&body, "ContributorInsightsAction")?;
-        let index_name = body["IndexName"].as_str();
-
-        let mut state = self.state.write();
-        let table = get_table_mut(&mut state.tables, table_name)?;
-
-        let status = match action {
-            "ENABLE" => "ENABLED",
-            "DISABLE" => "DISABLED",
-            _ => {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    format!("Invalid ContributorInsightsAction: {action}"),
-                ))
-            }
-        };
-        table.contributor_insights_status = status.to_string();
-        if status == "DISABLED" {
-            table.contributor_insights_counters.clear();
-        }
-
-        let mut result = json!({
-            "TableName": table_name,
-            "ContributorInsightsStatus": status
-        });
-        if let Some(idx) = index_name {
-            result["IndexName"] = json!(idx);
-        }
-
-        Self::ok_json(result)
-    }
-
-    fn list_contributor_insights(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length("tableName", body["TableName"].as_str(), 1, 1024)?;
-        validate_optional_range_i64("maxResults", body["MaxResults"].as_i64(), 0, 100)?;
-        let table_name = body["TableName"].as_str();
-
-        let state = self.state.read();
-        let summaries: Vec<Value> = state
-            .tables
-            .values()
-            .filter(|t| table_name.is_none() || table_name == Some(t.name.as_str()))
-            .map(|t| {
-                json!({
-                    "TableName": t.name,
-                    "ContributorInsightsStatus": t.contributor_insights_status
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "ContributorInsightsSummaries": summaries
-        }))
-    }
-
-    // ── Import/Export ──────────────────────────────────────────────────
-
-    fn export_table_to_point_in_time(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let table_arn = require_str(&body, "TableArn")?;
-        let s3_bucket = require_str(&body, "S3Bucket")?;
-        let s3_prefix = body["S3Prefix"].as_str();
-        let export_format = body["ExportFormat"].as_str().unwrap_or("DYNAMODB_JSON");
-
-        let state = self.state.read();
-        // Verify table exists and get items
-        let table = find_table_by_arn(&state.tables, table_arn)?;
-        let items = table.items.clone();
-        let item_count = items.len() as i64;
-
-        let now = Utc::now();
-        let export_arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}/export/{}",
-            state.region,
-            state.account_id,
-            table_arn.rsplit('/').next().unwrap_or("unknown"),
-            uuid::Uuid::new_v4()
-        );
-
-        drop(state);
-
-        // Serialize items as JSON Lines and write to S3
-        let mut json_lines = String::new();
-        for item in &items {
-            let item_json = if export_format == "DYNAMODB_JSON" {
-                json!({ "Item": item })
-            } else {
-                json!(item)
-            };
-            json_lines.push_str(&serde_json::to_string(&item_json).unwrap_or_default());
-            json_lines.push('\n');
-        }
-        let data_size = json_lines.len() as i64;
-
-        // Build S3 key for the export data
-        let s3_key = if let Some(prefix) = s3_prefix {
-            format!("{prefix}/data/manifest-files.json")
-        } else {
-            "data/manifest-files.json".to_string()
-        };
-
-        // Write to S3 if we have access to S3 state
-        let mut export_failed = false;
-        let mut failure_reason = String::new();
-        if let Some(ref s3_state) = self.s3_state {
-            let mut s3 = s3_state.write();
-            if let Some(bucket) = s3.buckets.get_mut(s3_bucket) {
-                let etag = uuid::Uuid::new_v4().to_string().replace('-', "");
-                let obj = fakecloud_s3::state::S3Object {
-                    key: s3_key.clone(),
-                    data: bytes::Bytes::from(json_lines),
-                    content_type: "application/json".to_string(),
-                    etag,
-                    size: data_size as u64,
-                    last_modified: now,
-                    metadata: HashMap::new(),
-                    storage_class: "STANDARD".to_string(),
-                    tags: HashMap::new(),
-                    acl_grants: Vec::new(),
-                    acl_owner_id: None,
-                    parts_count: None,
-                    part_sizes: None,
-                    sse_algorithm: None,
-                    sse_kms_key_id: None,
-                    bucket_key_enabled: None,
-                    version_id: None,
-                    is_delete_marker: false,
-                    content_encoding: None,
-                    website_redirect_location: None,
-                    restore_ongoing: None,
-                    restore_expiry: None,
-                    checksum_algorithm: None,
-                    checksum_value: None,
-                    lock_mode: None,
-                    lock_retain_until: None,
-                    lock_legal_hold: None,
-                };
-                bucket.objects.insert(s3_key, obj);
-            } else {
-                export_failed = true;
-                failure_reason = format!("S3 bucket does not exist: {s3_bucket}");
-            }
-        }
-
-        let export_status = if export_failed { "FAILED" } else { "COMPLETED" };
-
-        let export = ExportDescription {
-            export_arn: export_arn.clone(),
-            export_status: export_status.to_string(),
-            table_arn: table_arn.to_string(),
-            s3_bucket: s3_bucket.to_string(),
-            s3_prefix: s3_prefix.map(|s| s.to_string()),
-            export_format: export_format.to_string(),
-            start_time: now,
-            end_time: now,
-            export_time: now,
-            item_count,
-            billed_size_bytes: data_size,
-        };
-
-        let mut state = self.state.write();
-        state.exports.insert(export_arn.clone(), export);
-
-        let mut response = json!({
-            "ExportDescription": {
-                "ExportArn": export_arn,
-                "ExportStatus": export_status,
-                "TableArn": table_arn,
-                "S3Bucket": s3_bucket,
-                "S3Prefix": s3_prefix,
-                "ExportFormat": export_format,
-                "StartTime": now.timestamp() as f64,
-                "EndTime": now.timestamp() as f64,
-                "ExportTime": now.timestamp() as f64,
-                "ItemCount": item_count,
-                "BilledSizeBytes": data_size
-            }
-        });
-        if export_failed {
-            response["ExportDescription"]["FailureCode"] = json!("S3NoSuchBucket");
-            response["ExportDescription"]["FailureMessage"] = json!(failure_reason);
-        }
-        Self::ok_json(response)
-    }
-
-    fn describe_export(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let export_arn = require_str(&body, "ExportArn")?;
-
-        let state = self.state.read();
-        let export = state.exports.get(export_arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ExportNotFoundException",
-                format!("Export not found: {export_arn}"),
-            )
-        })?;
-
-        Self::ok_json(json!({
-            "ExportDescription": {
-                "ExportArn": export.export_arn,
-                "ExportStatus": export.export_status,
-                "TableArn": export.table_arn,
-                "S3Bucket": export.s3_bucket,
-                "S3Prefix": export.s3_prefix,
-                "ExportFormat": export.export_format,
-                "StartTime": export.start_time.timestamp() as f64,
-                "EndTime": export.end_time.timestamp() as f64,
-                "ExportTime": export.export_time.timestamp() as f64,
-                "ItemCount": export.item_count,
-                "BilledSizeBytes": export.billed_size_bytes
-            }
-        }))
-    }
-
-    fn list_exports(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length("tableArn", body["TableArn"].as_str(), 1, 1024)?;
-        validate_optional_range_i64("maxResults", body["MaxResults"].as_i64(), 1, 25)?;
-        let table_arn = body["TableArn"].as_str();
-
-        let state = self.state.read();
-        let summaries: Vec<Value> = state
-            .exports
-            .values()
-            .filter(|e| table_arn.is_none() || table_arn == Some(e.table_arn.as_str()))
-            .map(|e| {
-                json!({
-                    "ExportArn": e.export_arn,
-                    "ExportStatus": e.export_status,
-                    "TableArn": e.table_arn
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "ExportSummaries": summaries
-        }))
-    }
-
-    fn import_table(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let input_format = require_str(&body, "InputFormat")?;
-        let s3_source = body["S3BucketSource"].as_object().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "S3BucketSource is required",
-            )
-        })?;
-        let s3_bucket = s3_source
-            .get("S3Bucket")
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-        let s3_key_prefix = s3_source
-            .get("S3KeyPrefix")
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-
-        let table_params = body["TableCreationParameters"].as_object().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TableCreationParameters is required",
-            )
-        })?;
-        let table_name = table_params
-            .get("TableName")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "TableCreationParameters.TableName is required",
-                )
-            })?;
-
-        let key_schema = parse_key_schema(table_params.get("KeySchema").unwrap_or(&Value::Null))?;
-        let attribute_definitions = parse_attribute_definitions(
-            table_params
-                .get("AttributeDefinitions")
-                .unwrap_or(&Value::Null),
-        )?;
-
-        // Read items from S3 if we have access
-        let mut imported_items: Vec<HashMap<String, Value>> = Vec::new();
-        let mut processed_size_bytes: i64 = 0;
-        if let Some(ref s3_state) = self.s3_state {
-            let s3 = s3_state.read();
-            let bucket = s3.buckets.get(s3_bucket).ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ImportConflictException",
-                    format!("S3 bucket does not exist: {s3_bucket}"),
-                )
-            })?;
-            // Find all objects under the prefix and try to parse JSON Lines from each
-            let prefix = if s3_key_prefix.is_empty() {
-                String::new()
-            } else {
-                s3_key_prefix.to_string()
-            };
-            for (key, obj) in &bucket.objects {
-                if !prefix.is_empty() && !key.starts_with(&prefix) {
-                    continue;
-                }
-                let data = std::str::from_utf8(&obj.data).unwrap_or("");
-                processed_size_bytes += obj.size as i64;
-                for line in data.lines() {
-                    let line = line.trim();
-                    if line.is_empty() {
-                        continue;
-                    }
-                    if let Ok(parsed) = serde_json::from_str::<Value>(line) {
-                        // DYNAMODB_JSON format wraps items in {"Item": {...}}
-                        let item = if input_format == "DYNAMODB_JSON" {
-                            if let Some(item_obj) = parsed.get("Item") {
-                                item_obj.as_object().cloned().unwrap_or_default()
-                            } else {
-                                parsed.as_object().cloned().unwrap_or_default()
-                            }
-                        } else {
-                            parsed.as_object().cloned().unwrap_or_default()
-                        };
-                        if !item.is_empty() {
-                            imported_items
-                                .push(item.into_iter().collect::<HashMap<String, Value>>());
-                        }
-                    }
-                }
-            }
-        }
-
-        let mut state = self.state.write();
-
-        if state.tables.contains_key(table_name) {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ResourceInUseException",
-                format!("Table already exists: {table_name}"),
-            ));
-        }
-
-        let now = Utc::now();
-        let table_arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}",
-            state.region, state.account_id, table_name
-        );
-        let import_arn = format!(
-            "arn:aws:dynamodb:{}:{}:table/{}/import/{}",
-            state.region,
-            state.account_id,
-            table_name,
-            uuid::Uuid::new_v4()
-        );
-
-        let processed_item_count = imported_items.len() as i64;
-
-        let mut table = DynamoTable {
-            name: table_name.to_string(),
-            arn: table_arn.clone(),
-            key_schema,
-            attribute_definitions,
-            provisioned_throughput: ProvisionedThroughput {
-                read_capacity_units: 0,
-                write_capacity_units: 0,
-            },
-            items: imported_items,
-            gsi: Vec::new(),
-            lsi: Vec::new(),
-            tags: HashMap::new(),
-            created_at: now,
-            status: "ACTIVE".to_string(),
-            item_count: 0,
-            size_bytes: 0,
-            billing_mode: "PAY_PER_REQUEST".to_string(),
-            ttl_attribute: None,
-            ttl_enabled: false,
-            resource_policy: None,
-            pitr_enabled: false,
-            kinesis_destinations: Vec::new(),
-            contributor_insights_status: "DISABLED".to_string(),
-            contributor_insights_counters: HashMap::new(),
-            stream_enabled: false,
-            stream_view_type: None,
-            stream_arn: None,
-            stream_records: Arc::new(RwLock::new(Vec::new())),
-            sse_type: None,
-            sse_kms_key_arn: None,
-        };
-        table.recalculate_stats();
-        state.tables.insert(table_name.to_string(), table);
-
-        let import_desc = ImportDescription {
-            import_arn: import_arn.clone(),
-            import_status: "COMPLETED".to_string(),
-            table_arn: table_arn.clone(),
-            table_name: table_name.to_string(),
-            s3_bucket_source: s3_bucket.to_string(),
-            input_format: input_format.to_string(),
-            start_time: now,
-            end_time: now,
-            processed_item_count,
-            processed_size_bytes,
-        };
-        state.imports.insert(import_arn.clone(), import_desc);
-
-        let table_ref = state.tables.get(table_name).unwrap();
-        let ks: Vec<Value> = table_ref
-            .key_schema
-            .iter()
-            .map(|k| json!({"AttributeName": k.attribute_name, "KeyType": k.key_type}))
-            .collect();
-        let ad: Vec<Value> = table_ref
-            .attribute_definitions
-            .iter()
-            .map(|a| json!({"AttributeName": a.attribute_name, "AttributeType": a.attribute_type}))
-            .collect();
-
-        Self::ok_json(json!({
-            "ImportTableDescription": {
-                "ImportArn": import_arn,
-                "ImportStatus": "COMPLETED",
-                "TableArn": table_arn,
-                "TableId": uuid::Uuid::new_v4().to_string(),
-                "S3BucketSource": {
-                    "S3Bucket": s3_bucket
-                },
-                "InputFormat": input_format,
-                "TableCreationParameters": {
-                    "TableName": table_name,
-                    "KeySchema": ks,
-                    "AttributeDefinitions": ad
-                },
-                "StartTime": now.timestamp() as f64,
-                "EndTime": now.timestamp() as f64,
-                "ProcessedItemCount": processed_item_count,
-                "ProcessedSizeBytes": processed_size_bytes
-            }
-        }))
-    }
-
-    fn describe_import(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        let import_arn = require_str(&body, "ImportArn")?;
-
-        let state = self.state.read();
-        let import = state.imports.get(import_arn).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ImportNotFoundException",
-                format!("Import not found: {import_arn}"),
-            )
-        })?;
-
-        Self::ok_json(json!({
-            "ImportTableDescription": {
-                "ImportArn": import.import_arn,
-                "ImportStatus": import.import_status,
-                "TableArn": import.table_arn,
-                "S3BucketSource": {
-                    "S3Bucket": import.s3_bucket_source
-                },
-                "InputFormat": import.input_format,
-                "StartTime": import.start_time.timestamp() as f64,
-                "EndTime": import.end_time.timestamp() as f64,
-                "ProcessedItemCount": import.processed_item_count,
-                "ProcessedSizeBytes": import.processed_size_bytes
-            }
-        }))
-    }
-
-    fn list_imports(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = Self::parse_body(req)?;
-        validate_optional_string_length("tableArn", body["TableArn"].as_str(), 1, 1024)?;
-        validate_optional_string_length("nextToken", body["NextToken"].as_str(), 112, 1024)?;
-        validate_optional_range_i64("pageSize", body["PageSize"].as_i64(), 1, 25)?;
-        let table_arn = body["TableArn"].as_str();
-
-        let state = self.state.read();
-        let summaries: Vec<Value> = state
-            .imports
-            .values()
-            .filter(|i| table_arn.is_none() || table_arn == Some(i.table_arn.as_str()))
-            .map(|i| {
-                json!({
-                    "ImportArn": i.import_arn,
-                    "ImportStatus": i.import_status,
-                    "TableArn": i.table_arn
-                })
-            })
-            .collect();
-
-        Self::ok_json(json!({
-            "ImportSummaryList": summaries
-        }))
     }
 }
 
@@ -3369,7 +264,6 @@ impl AwsService for DynamoDbService {
         ]
     }
 }
-
 // ── Helper functions ────────────────────────────────────────────────────
 
 fn require_str<'a>(body: &'a Value, field: &str) -> Result<&'a str, AwsServiceError> {
